+++ LookupAnything/Framework/Fields/FishSpawnRulesField.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Xna.Framework;
+using Pathoschild.Stardew.Common;
 using Pathoschild.Stardew.LookupAnything.Framework.Models.FishData;
 using StardewValley;
 using StardewValley.ItemTypeDefinitions;
@@ -42,8 +43,8 @@ internal class FishSpawnRulesField : CheckboxListField
     private IEnumerable<KeyValuePair<IFormattedText[], bool>> GetConditions(GameHelper gameHelper, ParsedItemData fish)
     {
         // get spawn data
-        FishSpawnData? spawnRules = gameHelper.GetFishSpawnRules(fish);
-        if (spawnRules?.Locations?.Any() != true)
+        FishSpawnData spawnRules = gameHelper.GetFishSpawnRules(fish);
+        if (spawnRules.Locations?.Any() != true)
             yield break;
 
         // not caught uet
@@ -70,7 +71,7 @@ internal class FishSpawnRulesField : CheckboxListField
             yield return this.GetCondition(
                 label: I18n.Item_FishSpawnRules_Time(
                     times: I18n.List(
-                        spawnRules.TimesOfDay.Select(p => I18n.Generic_Range(Game1.getTimeOfDayString(p.MinTime), Game1.getTimeOfDayString(p.MaxTime)).ToString())
+                        spawnRules.TimesOfDay.Select(p => I18n.Generic_Range(CommonHelper.FormatTime(p.MinTime), CommonHelper.FormatTime(p.MaxTime)).ToString())
                     )
                 ),
                 isMet: spawnRules.TimesOfDay.Any(p => Game1.timeOfDay >= p.MinTime && Game1.timeOfDay <= p.MaxTime)